This merge adds four new features to the ocean testing infrastructure.

To begin with, it adds the ability to generate a driver script for an
entire test case. This can automate the process of running multiple
steps (or analyzing output).

In addition, it fixes issues related to generating multiple namelist and
streams files within each case directory.

It also adds the ability to use pre-defined executables in tags used
within driver and run scripts.

Finally, it modifies the way files are downloaded. Now the tag describes
where a file will be downloaded into, and the tag is used to link (or
copy) a file into a case directory. This allows a single file to be used
in multiple cases without needing copies in each case (unless that is
requested).

In addition, it performs some cleanup on the 642 basic spherical test
case and adds a .gitignore file to prevent generated files from
accidentally being committed.

* testing/general_enhancements:
  Allow the use of pre-defined executables in case configs
  Updating <get_file> tag to download into a backup location
  Adding driver script generation
  Add generation of multiple namelist / streams files
  Cleaning up the 642 basic spherical test

+++ test_cases/ocean/clean_testcase.py
@@ -12,7 +12,6 @@ parser.add_argument("-r", "--resolution", dest="resolution", help="Resolution of
 
 args = parser.parse_args()
 
-
 base_path = '%s/%s/%s'%(args.core, args.configuration, args.resolution)
 for file in os.listdir('%s'%(base_path)):
 	if fnmatch.fnmatch(file, '*.xml'):
@@ -21,11 +20,20 @@ for file in os.listdir('%s'%(base_path)):
 		config_tree = ET.parse(config_file)
 		config_root = config_tree.getroot()
 
-		case_dir = config_root.attrib['case']
+		if config_root.tag == 'config':
+			case_dir = config_root.attrib['case']
+
+
+			if os.path.exists('%s/%s'%(base_path, case_dir)):
+				if os.path.isdir('%s/%s'%(base_path, case_dir)):
+					shutil.rmtree('%s/%s'%(base_path, case_dir))
+					print ' -- Removed case %s/%s'%(base_path, case_dir)
+		elif config_root.tag == 'driver_script':
+			script_name = config_root.attrib['name']
+
+			if os.path.exists('%s/%s'%(base_path, script_name)):
+				os.remove('%s/%s'%(base_path, script_name))
+				print ' -- Removed driver script %s/%s'%(base_path, script_name)
 
 		del config_tree
 		del config_root
-
-		if os.path.exists('%s/%s'%(base_path, case_dir)):
-			shutil.rmtree('%s/%s'%(base_path, case_dir))
-			print " -- Removed case %s/%s"%(base_path, case_dir)
